[fix] QA-040반영 :  DateFormat추가 및 적용
