+++ BongBaek/BongBaek/Presentation/Home/Schedule/ScheduleAlarmCellView.swift
@@ -54,7 +54,7 @@ struct ScheduleIndicatorCellView: View {
                     .frame(width: 14, height: 14)
                     .padding(.leading, 8)
                 
-                Text(formatDate(event.eventInfo.eventDate))
+                Text(event.eventInfo.eventDate.DateFormat())
                     .font(.caption_regular_12)
                     .foregroundColor(.gray100)
             }
@@ -83,18 +83,4 @@ struct ScheduleIndicatorCellView: View {
         .frame(height: 260)
         .cornerRadius(10)
     }
-}
-
-public func formatDate(_ dateString: String) -> String {
-    let inputFormatter = DateFormatter()
-    inputFormatter.dateFormat = "yyyy-MM-dd"
-    
-    let outputFormatter = DateFormatter()
-    outputFormatter.dateFormat = "yyyy년 MM월 dd일"
-    
-    if let date = inputFormatter.date(from: dateString) {
-        return outputFormatter.string(from: date)
-    }
-    
-    return dateString // 변환 실패 시 원본 반환
 }
